ts-web/core: move OasisNodeClient to client.NodeInternal
It's anticipated that we'll add typed wrappers for consensus transactions,
which will be a layer above the gRPC wrapper. I think that it will be better
to move this out of the index module if something even higher will depend on
it.
